# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ajosave API
  description: |
    Trustless rotating savings circles (Ajo/Esusu) on the Stellar blockchain.

    ## API Versioning
    This API uses URL-based versioning. All endpoints are prefixed with `/api/v1/`.
    
    **Current version:** v1  
    **Base URL:** `/api/v1/`

    Legacy endpoints without version prefix (`/api/`) are deprecated and redirect to v1.

    ## Authentication
    Most endpoints require a valid session cookie obtained via the OTP login flow:
    1. `POST /api/v1/auth/send-otp` – request a 6-digit OTP to your phone
    2. `POST /api/v1/auth/verify-otp` – exchange OTP for a session cookie (NextAuth)

    The session cookie is automatically sent by the browser on subsequent requests.
    Server-to-server callers must include the `Cookie` header.

    The cron endpoint uses a separate `Authorization: Bearer <CRON_SECRET>` header.
  version: "1.0.0"
  contact:
    email: security@ajosave.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://www.ajosave.app/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie set after OTP verification
    cronBearer:
      type: http
      scheme: bearer
      description: Static secret from `CRON_SECRET` env var
  schemas:
...
    ApiSuccess:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          description: Payload – see individual endpoint schemas
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          example: "Circle not found"
        code:
          type: string
          example: "UNAUTHORIZED"

    # ── Domain objects ──────────────────────────────────────────────────────────

    # Soroban AjoContract.get_state() returns a 5-tuple:
    #   (current_cycle: u32, max_members: u32, next_payout_time: u64, completed: bool, paused: bool)
    # This schema mirrors that tuple for any REST endpoint that surfaces on-chain state.
...
  responses:
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            success: false
            error: "Unauthorized"
            code: "UNAUTHORIZED"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            success: false
            error: "Circle not found"
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            success: false
            error: "Internal server error"
            code: "INTERNAL_ERROR"

# ─── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /auth/logout:
    post:
      tags: [auth]
      summary: Sign out
      description: Clears the NextAuth session cookie.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Signed out successfully

  # ── Circles ─────────────────────────────────────────────────────────────────
...
  /circles/{id}/contribute:
    post:
      tags: [circles]
      summary: Initiate a contribution payment
      description: |
        Creates a Paystack payment session for the current cycle contribution.
        Redirect the user to `authorizationUrl` to complete payment.
        After payment, Paystack calls the verify webhook to confirm on-chain.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Paystack authorization URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [authorizationUrl, reference]
                        properties:
                          authorizationUrl:
                            type: string
                            format: uri
                            example: "https://checkout.paystack.com/abc123"
                          reference:
                            type: string
                            example: "ajo-circle-uuid-1-a1b2c3d4"
        "400":
          description: Circle not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not a member of this circle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Cron ────────────────────────────────────────────────────────────────────
...